+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrelesead] - 2021-01-29
+### Changed
+- Fix __str__ di Stamping per i casi in cui alcuni attributi non sono presenti 
+
 ## [1.2.2] - 2021-01-15
 ### Changed
 - Fix configurazione cron per compatibilità con Debian (immagine docker python:3-slim)